Cut down on the massive memory consumption
Previously all the lines ever comitted to a repository (that match any
of the regular expressions) would stay in memory for the entire runtime
of gittyleaks. Now only the hashes of the lines of the current revision
stay in memory as that is enough to avoid reporting repetitive lines
multiple times further down the chain.

You could argue that we can remove the cache entirely as the work done
here will be done in a later stage anyway.

As a side-effect of this change `get_word_matches` is now a generator in
constrast to returning a list. This allowed getting rid of temporary
buffer set that contained all the strings. The results will now be
passed down the chain and will be removed from memory if they aren't
interesting at all.

I experimented with having `get_git_matches` yield lines as they are
piped in from git but that didn't yield any improvement in runtime or
memory over all. It did cut down on the initial memory spiky tho.. As
every single `git grep` invocation returned about 75MiB of text. In the
grand scheme of scanning the entire repository on every commit 75MiB
isn't important at all. The primary bottleneck is still `git grep`'s
runtime performance.

Another change I did was ignoring any kind of unicode decode errors.
Those characters will now be skipped. This isn't ideal but at least
gittyleaks doesn't crap out on them.